Moved PVSurface and PVSegment to base.py + moved pvground.py to geom
* and new config params
