+++ pvfactors/config.py
@@ -2,3 +2,9 @@
 
 DEFAULT_NORMAL_VEC = None
 TOL_COLLINEAR = 1e-5
+
+
+# Ground params
+MAX_X_GROUND = 1e2
+MIN_X_GROUND = - MAX_X_GROUND
+Y_GROUND = 0.
