+++ pvfactors/geometry/__init__.py
@@ -1,6 +1,7 @@
 """Geometry classes"""
 
 from pvfactors.geometry.base import \
-    BaseSide, ShadeCollection, BaseSurface
+    BaseSide, ShadeCollection, BaseSurface, PVSurface, PVSegment
 from pvfactors.geometry.pvrow import PVRowSide, PVRow
-from pvfactors.geometry.pvsurface import PVSurface, PVSegment
+from pvfactors.geometry.pvarray import OrderedPVArray
+from pvfactors.geometry.pvground import PVGround
